Refactor renderAsd function into a separate module
Moved the renderAsd function from extension.ts to a new file renderAsd.ts for better modularity and code organization. This helps in maintaining and testing the function independently from the main codebase. Updated import statements accordingly to reflect the change.
